requirement.txt
langchain
python-dotenv
langchain-community
beautifulsoup4
langgraph
langchain-core
langchain-text-splitters
chromadb
faiss-cpu
langchain-chroma
streamlit
langchain-groq
reportlab
pandas
requests
matplotlib